Boogu: consolidate model into a single file (single-file convention)
diffusers follows a one-model-one-file convention. Merge the Boogu model's
helper modules into transformer_boogu.py:
  - rope_boogu.py            -> RoPE section
  - block_lumina2.py         -> norm / feed-forward / embedding section
  - attention_processor_boogu.py -> attention-processor section
Update the two pipelines and the transformer test to import
BooguImageRotaryPosEmbed from transformer_boogu. Pure code relocation: the
class bodies are unchanged, so checkpoints load identically and base/edit/turbo
remain bit-exact (verified end-to-end on GPU). Addresses reviewer single-file
convention feedback.
